Tightens up the RCON settings reconciliation loop
Here, we apply desired settings via RCON, using the response line from the
server to update the observed value.  This means we don't need to re-read all
the settings from the server again.  Also, the parsing code updates the existing
structs instead of returning new ones that may have incorrect default values.

There's still an issue with this version when you don't provide a desired value
in the server spec; in that case, we end up in an infinite loop because we don't
know what the default value is supposed to be.