Solves the periodic settings reconciliation without compounding requeues
Using the approach described in the article _[Implementing a periodic Kubernetes
controller][1]_, this sets up an event channel that is monitoring for servers
that haven't had a reconciliation in 10 seconds and requeues them.  This allows
for a much less noisy reconciliation process, since there is no compounding
requeuing with `RequeueAfter`.

[1]: https://dnaeon.github.io/kubernetes-periodic-controller/